feat: allow for override of module output via config
This way, if the user wants to change how a specific module displays
its information to the user, they can simply manipulate the `output`
option in their config file. This is in line with how we determine the
output for the prompt module -- the user simply sets `output` under
`[modules.prompt]` to be whatever they want, and it will get printed
instead of the default values.

Note: due to the way bash's `PS1` variable is evaluated, doing
something like `output = "$(pwd)"` will actually be evaluated; it will
not literally say "$(pwd)". So don't put anything stupid like "$(rm
-rf ~/*)" inside of an `output` option.