feat: allow for passing of text style options
Crate `ansi-term` allows for the passing of various text style options
to format the text with. The `bold` property will make the text bold,
the `italic` property will italicize the text, etc...

We now allow for the passing of a `style` option (global or by-module)
that will be converted into the appropriate `ansi_term::Style` object.