feat: conversion method from string to ansi_term::Color
The user should be able to specify what color to make the
foreground/background of their modules in a config file. Therefore, we
need conversion methods that can take in some kind of input (in this
case, a string) and convert it to the appropriate `ansi_term::Color`.

With the `try_color_from_str()` method, we can get a string
representing a color from the user's config file, such as "blue", and
convert it into `ansi_term::Color::Blue`. The method will return a
`FromStringError` if the input doesn't correspond to a valid `Color`.