feat: conversion methods for ModuleStyle to ansi_term::Style
We choose to hold the options that each module will always have within
a struct called `ModuleOptions`. Within that struct, we have another
struct (`ModuleStyle`) that contains information about the background
color and foreground color of the text that will be formatted by the
`format_for_module()` method.

We can now convert our struct into an `ansi_term::Style` that can be
used to format strings.