feat: implement format_for_module
At their core, each module basically does one thing and the same: they
take in a string and options on how to style it, and they output
another string that can be printed by a terminal emulator to reflect
that style. The "generic" method `format_for_module()` can handle this
basic requirement, plus it's much easier to test as its own method.