feat: implement try_fixed_from_str
We can now convert strings to `ansi_term::Color::Fixed` via the
`try_fixed_from_str()` method. As long as the input can be coerced
into a `u8`, we can create the corresponding `Color`. Malformed inputs
will return an error.

At this point, we should now be able to take in any user input and
convert it into the correct `Color`.