feat: ability to parse module styles from configs
With the addition of many new helper methods, we can now take in a
config file and a key representing the name of the module whose style
we're trying to find, and convert the user's options into a
`ModuleStyle` object understood by the rest of the program. The user
can represent colors as words, tuples, or integers -- if it parses
correctly, it gets used. They can also set either a single text
property to apply to the text, or multiple if they so choose.

Below is what kind of config file the program expects:
```toml
[modules.<module_name>]

[modules.<module_name>.style]
foreground = "green"        # use of a word
background = "(0, 60, 65)"  # use of an RGB tuple
text_properties = "italic"  # array of strings is also valid here
```