feat: implement parsing and printing of modules by name
The program will now read the config file, format the appropriate
modules, and print them to the terminal. If a module is specified in
the config file but doesn't correspond to a built-in module, the
program will attempt to find the style and output of said module
within the config file, and use that information to print it.