refactor: reimplement ModuleError as Error
The original `ModuleError` type was lacking with regards to providing
the user (and the programmer) with helpful information whenever the
program crashed. Errors in the config file were all masked behind
`ModuleError::InvalidForm`, and it was impossible to know what exactly
had gone wrong. Putting a `u8` where a `String` was expected yielded
the same opaque message as making a mistake when specifying an RGB u8
tuple, and the messages didn't include what had gotten messed up, only
that an error had been encountered.

The new `Error` type is based off of the implementation in the `git2`
crate. Now, whenever the user does something the program doesn't
expect (like inputting an invalid type into their config), we actually
differentiate between the different kinds of `Error` via the
`ErrorKind` enum, and we display a message that explains what went
wrong and what we expected. For example, if the user specifies that
the `foreground` of the `cwd` module is equal to `-1` (invalid,
because only strings and `u8`s are allowed), we get the following
error:

```
...panicked at 'InvalidTypeInConfig: expected u8, got: -1'...
```

And if the user tries to specify a color that doesn't exist (like
"turquoise"):

```
...panicked at 'ConfigParseFailure: expected valid color, u8, or rgb tuple, got: "turquoise"'...
```

(Note that the above error should really be a `NoSuchMatchInConfig`,
but in this specific case we can't differentiate between the two
errors because color u8 parsing and color rgb parsing can also fail
for different reasons.)