Add basic config file retrieval
Retrieves the main config file from
"$XDG_CONFIG_HOME/contrail/config.toml" and the module config files from
"$XDG_CONFIG_HOME/contrail/modules/" if there are any. Panics with an
error message if either the CLI-specified folder path or the default one
doesn't exist. In the future, it should not be an error for the default
config file to not exist; we should just use the program defaults.

Resolves: #24