Implement multithreaded command execution
Commands are gathered via clap and executed concurrently. Whatever the
commands print to stdout is then concatenated (in the order that they
are passed to contrail) and printed back to stdout. No newlines are
stripped, so everything will likely print on its own line.

Currently, there is no way to handle flags and arguments passed to the
commands fed into contrail.