Flush stdout after print; Cargo Update
I had been noticing a couple of situations, especially right after
opening a shell, where it takes some time to get a shell. I know that
contrail doesn't take that long to run, so I added stdout().flush() so
that bash no longer gets confused after contrail exits. This is a
significant improvement in responsiveness.

Also there was a build error, so I ran cargo update. IDK if that was an
appropriate thing to do, but there it is.